WIP to remove function keyword from closures.
Since they would be mixed with statements and written "unsized_type name(args)", the unsized_type will collide with decls' sized_type.
To avoid that collision, you can unify unsized and sized types into one rule, basic_type, and have it return a Type.t to be cast later. That solves the basic types.

The other problem comes in the collision of "array" keywords with the LHS of assignments.
We can't use ARRAY in unsized_type since it can be an identifier and that makes it a shift/reduce with LHS.
So, we have to match "array" with LHS. But we can't just match "array" since then the LHS "array[,,,]" would clash with the LHS "array" + the unsized dimensions "[,,,]". So, we have to match the whole "array[,,,]" as a LHS.

Incomplete:
 - Migrate top_var_types to top_var_extension (that's where the errors about unreachable constraint rules come from)
 - Fix the types by making the other, non-basic types return Type.t; and also wrap those in Type.to_sized and Type.to_unsized (or rather roll your own for better errors, since they're user facing)
 - Fix error message :( hopefully not too bad with the update tool